+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 0.5.8
 
-* Add `formatOutput` optional parameter to the `GeneratorBuilder` constructor,
-  which defaults to `true`. Setting this to false will speed up builds at the
-  cost of uglier output.
+* Add `formatOutput` optional parameter to the `GeneratorBuilder` constructor.
+  This is a lamda of the form `String formatOutput(String originalCode)` which
+  allows you do do custom formatting. If a `null` value is passed then no
+  formatting will be done.
 
 ## 0.5.7